requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
google-generativeai==0.3.1
pytest==7.4.3
httpx==0.25.1
python-multipart==0.0.6
sympy==1.12
numpy==1.26.1
aiohttp==3.8.6
google-adk==0.1.0

# MongoDB dependencies
motor==3.3.2
beanie==1.23.0
pymongo==4.6.1
